Swap to TypeScript
This doesn't migrate any of the source files, but it convinces all the
tooling to start operating with TypeScript. It should be relatively easy
to migrate since there's hardly any code.

EDIT: there were no required changes.

I did this because `media-devices` is already in TypeScript and I'd very
much like to document my redux store structure in a typed language.