Add basic action to get media devices
Whew, this was a surprising amount of work. I always forget how much
there is to slog through in a new app. Most of the work was setting up
testing conventions, mocks, and fixing the linter (which didn't quite
support TypeScript syntax).

The new action doesn't do anything interesting yet, and I'll probably
change a fair bit of it, but this gives me something to start with.