Add media tracks to state & global context
Redux is starting to take shape. I've created the notion of participants
and tracks, with the real track instances kept in a global state
singleton. That's an idea I pioneered in another project called
"gather", but it never reached the light of day, although other
experiments have proven successful so I'll use the same idea here.

Now that redux and global context are in alignment, I can start working
on a video display component. Hooray, progress!