Render media streams as they become available
This binds the `MediaView` component to a set of tracks and a media
stream. As you update the audio/video tracks, the media stream updates
in turn.

This could come back to bite me if I re-arrange the video layout. It's
possible that the media stream and video may need to persist on the page
to avoid audio gaps. That requires more manual testing.