Replace one-shot message pipeline with a queue
If you use `pipe([...])` to send a single message, `it-pipe` notices the
stream is bounded and automatically closes the iterator. As
a consequence, the stream (and connection) is closed.

I've replaced the message source with an asynchronous queue with an
imperative interface. Because the message source no longer closes after
the first message, the stream remains open until manually closed.