Implement high-level messenger for data channels
This was going to be used to send a "close" event to the remote peer,
but then I found my old comment:

> Permanently closes the peer connection. The 'close' event fires on
> the remote data channel.

Sure enough, it worked. Whatever. This was the next step after getting
close events working, and I was already halfway through.

It's nice working on personal projects. I can leave work half-finished
and commit straight to master. Nobody cares.