Unify signaling server with Farnsworth repo
I'd been keeping the signaling server in a separate repo, but
considering how intertwined they are, it seemed reasonable to combine
them.

The other repo was more of a proof of concept (read: absolute trash) so
it took some doing to raise the code quality. There's still a bit more
to do regarding repo setup and config generation, but this is good for
now.