Define utility for inspecting application state
This provides a way to access the conferencing state in production. It
is frequently useful to pull a peer connection and run some manual
commands, but it's much more difficult if you're in production (i.e.
when you really *need* to know what went wrong).

Call `tinker()` to get access to the debugging toolkit.