Remove invite code input form
The invite form used to accept a peer ID as a dial address. It was
mainly used in testing. Now that I have a call route, I don't need it
anymore, nor do I expect other people to find value in it. Well, it
might be useful if they have different domains for custom builds of the
web app. But that seems more distant than is useful to practical to plan
for.