Prevent pure whitespace messages
Okay, you could still technically do it through the console, but this
prevents it at the UI level. I'm not terribly worried about it in other
cases. If someone is hacking around those constraints, they're probably
having fun, and the person they're harassing is literally on the phone
with them. The scope for actual trouble seems limited.