Tell the other participant when you share a screen
I may end up reverting this, but it gives me a good stopping point. I'm
going to work on something else for a bit. Wayland is giving me trouble
with screen sharing and using Chromium to work around it is slowly
draining my soul.

This currently suffers from an ordering problem: the tracks arrive after
the event. While it's technically possible for Redux to handle that
ordering, it leaves space open for very confusing states. I'd rather
avoid diverging states if possible.