WIP: Clear state and run GC after leaving call
This implements another garbage collector handle I'd been planning:
dereferenced connection IDs. If you clear the connection ID out of
`state.participants`, the garbage collector will notice and
automatically close and delete the dereferenced WebRTC connection.

This is marked WIP because closing the PC triggers a data channel event
that perceives it as a *remote* close event, which tries to delete the
connection again and fails.

I need to mute that second event.