Close the libp2p stream when the call ends
If one of you leaves intentionally (e.g. refresh or end call button),
the application will close the libp2p signaling stream. It doesn't close
the entire connection, because it seems there isn't a good handle for
that.

I might add a hook in the garbage collector later to wipe out the entire
connection rather than just a stream, but today does not need to be that
day.