Use simpler ID for local participant state
This is easier to inspect and understand while debugging the
application. As a bonus, I don't have to copy+paste IDs in order to
index state.

At one point I planned on using the local libp2p identifier, but because
it's initialized asynchronously and capable of failure, I later decided
against it.