Detect and remove remotely ended tracks
This commit detects track disassociation from it's corresponding remote
stream to mean it ended. We can't close the RTP transceiver (triggering
a *true* 'ended' event) without also closing the associated inbound
track, but what we can do is remove the track from the sender which
triggers that `onremovetrack` event on all associated streams.

It's not ideal, but it seems to be the most reliable method short of
sending a data channel message. And I'm not stoked about managing
session metadata over the data channel.

Fixes #40