Add device and group IDs to track state
This changed a lot of tests. Since I anticipate adding more fields to
track state in the future, I've created a track factory and replaced all
the broken types with the wrapper. It provides sane defaults for track
state and allows you to override the bits you care about (a common
pattern in this codebase). That should help avoid these large changes
every time I need to add a field.