Replace existing tracks when changing devices
All that's necessary is cleaning out the reference to the old track. The
middleware handles garbage collection, bubbling through to the other
end, and the existing logic to send new tracks works just fine as is.
I just tested this with my other computer. It freaking works!

I can't believe how easy that was.