Keep track of camera facing modes
Each track now has a "facing mode", which corresponds to a setting on
locally generated video tracks if you're on a mobile device. It's pretty
specific. The goal is to disable video mirroring if you're showing your
environment.

I'm not a fan of putting all these nullable fields on every track, but
I don't anticipate this being a problem until many more fields are
added. Just thinking aloud right now.