Support credentials in default ICE servers
This is a breaking change for deployments: it requires switching
`STUN_SERVERS` to `ICE_SERVERS` and prefixing with the appropriate
protocol type.

This allows other people to compile in TURN support with user/pass
credentials, or OAuth tokens if they're super fancy.

I plan to keep my own TURN server private, but it just made more sense
to build it with that capability. Easier to deal with stun servers if
they all follow the same convention.