Ignore duplicate tracks
This is part of an effort to move local track sending to a middleware
function. After every dispatch, I want to send the locally-generated
tracks to each remote peer, but that requires checking if the track
already exists, so this commit ignores the track if it's already in the
senders list.